+++ plugins/storage/redis/src/hopeit/redis_storage/__init__.py
@@ -2,7 +2,7 @@
 Storage/persistence asynchronous get/set key-values.
 Backed by Redis
 """
-from typing import Optional, Type, Generic, Any
+from typing import Optional, Type, Generic, Any, List
 
 import aioredis
 
@@ -18,7 +18,7 @@ class RedisStorage(Generic[DataObject]):
        This class must be initialized with the method connect
        Example:
            ```
-           redis_store = await RedisStorage().connect(address="redis://hostname:6379")
+           redis_store = RedisStorage().connect(address="redis://hostname:6379")
            ```
     """
     def __init__(self):
@@ -50,13 +50,49 @@ class RedisStorage(Generic[DataObject]):
             return Payload.from_json(payload_str, datatype)
         return None
 
-    async def store(self, key: str, value: DataObject):
+    async def store(self, key: str, value: DataObject, **kwargs):
         """
         Stores value under specified key
 
         :param key: str
         :param value: DataObject, instance of dataclass annotated with @dataobject
+        :param **kwargs: You can use arguments expected by the set method in the aioredis library i.e.:
+            ex sets an expire flag on key name for ex seconds.
+            px sets an expire flag on key name for px milliseconds.
+            nx if set to True, set the value at key name to value only if it does not exist.
+            xx if set to True, set the value at key name to value only if it already exists.
+            keepttl if True, retain the time to live associated with the key. (Available since Redis 6.0).
+            *These arguments may vary depending on the version of aioredis installed.
+
+        i.e. store object:
+        ```
+        redis_store.store(key='my_key', value=my_dataobject)
+        ```
+
+        i.e. store object with kwargs option, adding `ex=60` redis set a ttl of 60 seconds for the object:
+        ```
+        redis_store.store(key='my_key', value=my_dataobject, ex=60)
+        ```
+
         """
         assert self._conn
         payload_str = str(Payload.to_json(value))
-        await self._conn.set(key, payload_str)
+        await self._conn.set(key, payload_str, **kwargs)
+
+    async def delete(self, *keys: str):
+        """
+        Delete specified keys
+
+        :param keys: str, keys to be deleted
+        """
+        assert self._conn
+        await self._conn.delete(*keys)
+
+    async def list_objects(self, wildcard: str = "*") -> List[str]:
+        """
+        Returns a list of keys matching `wildcard`
+
+        :param wildcard: str, expected glob-style wildcards
+        """
+        assert self._conn
+        return [obj.decode('utf-8') for obj in await self._conn.keys(wildcard)]